#37714 Improved activity stream note creation in site ctx
This adds logic to the activity stream widget and note input widget to better handle the state when no project is available in the context (e.g. a "site" context).

Notes that are created require a project to be set and there are some cases when a project isn't available. Normally, we pull the project from the context but in the site context situation this is not an option. Instead, the note creation widget now tries to look for a project on the entity itself and if that fails in errors. The activity stream in turn hides the note input widgets in these cases.

Closes #32