#38223 No duplicate notes in activity stream
Fixes an issue where previously, if you replied to a note in the activity stream, you would see the note twice - once right at the top with the new reply and once further down without the reply. (it was only a temporary state, a UI refresh would make it go way). This change hides the note further down in the list, so that you only see the note once, first in the activity stream.

Closes #48