#38753 Makes it possible to use field delegate without SG model
Adds a data_role property to the delegate that defaults to the current
data role from the SG model but allows you to set it to anything. These
changes maintain the current behavior but also account for usage of non-SG
models. Because SG models do some special things with the first column in
the model and the delegate can use that for efficiency sake, the delegate
will now inspect the indexes being processed to determine if the source model
is a SG model or not and behave accordingly.

I'm not 100% sure this is the best solution here, but it seems to work. I've
created a demo for this in the demo app that shows its usage.

https://github.com/shotgunsoftware/tk-multi-demo/blob/master/python/tk_multi_demo/demos/field_widget_delegate/demo.py#L109-L151

Will be great to get some other eyes on this.