SG-10692: Implements a placeholder thumbnail detection routine and cache refresh when required. (#91)
Recent changes around how thumbnails behave when they are uploaded to Shotgun has resulted in a caching-related problem in the activity stream widget. The widget maintains a persistent sqlite cache, and includes the URLs for any thumbnails associated with items in the activity stream. If the cache contained a URL to a placeholder thumbnail, it would never be purged and the user would always see the placeholder instead of the "hero" thumbnail.

The fix is to detect the placeholder thumbnail when it's downloaded. When we do, we trigger a single-shot QTimer with a 20-second interval that dumps the cache and re-queries the activity stream data from Shotgun.